+++ trunk-recorder/plugin_manager/plugin_manager.h
@@ -31,7 +31,7 @@ void start_plugins(std::vector<Source *> sources, std::vector<System *> systems)
 void stop_plugins();
 
 void plugman_poll_one();
-void plugman_audio_callback(Recorder *recorder, float *samples, int sampleCount);
+void plugman_audio_callback(Recorder *recorder, int16_t *samples, int sampleCount);
 int plugman_signal(long unitId, const char *signaling_type, gr::blocks::SignalType sig_type, Call *call, System *system, Recorder *recorder);
 int plugman_call_start(Call *call);
 int plugman_call_end(Call_Data_t call_info);
